Enum diem_framework_releases::legacy::transaction_scripts::LegacyStdlibScript

pub enum LegacyStdlibScript {

    AddCurrencyToAccount,
    AddRecoveryRotationCapability,
    AddValidatorAndReconfigure,
    Burn,
    BurnTxnFees,
    CancelBurn,
    CreateChildVaspAccount,
    CreateDesignatedDealer,
    CreateParentVaspAccount,
    CreateRecoveryAddress,
    CreateValidatorAccount,
    CreateValidatorOperatorAccount,
    FreezeAccount,
    PeerToPeerWithMetadata,
    Preburn,
    PublishSharedEd2551PublicKey,
    RegisterValidatorConfig,
    RemoveValidatorAndReconfigure,
    RotateAuthenticationKey,
    RotateAuthenticationKeyWithNonce,
    RotateAuthenticationKeyWithNonceAdmin,
    RotateAuthenticationKeyWithRecoveryAddress,
    RotateDualAttestationInfo,
    RotateSharedEd2551PublicKey,
    SetValidatorConfigAndReconfigure,
    SetValidatorOperator,
    SetValidatorOperatorWithNonceAdmin,
    TieredMint,
    UnfreezeAccount,
    UpdateExchangeRate,
    UpdateDiemVersion,
    UpdateMintingAbility,
    UpdateDualAttestationLimit,

}

All of the Move transaction scripts that can be executed on the Diem blockchain

Variants

AddCurrencyToAccount

AddRecoveryRotationCapability

AddValidatorAndReconfigure

Burn

BurnTxnFees

CancelBurn

CreateChildVaspAccount

CreateDesignatedDealer

CreateParentVaspAccount

CreateRecoveryAddress

CreateValidatorAccount

CreateValidatorOperatorAccount

FreezeAccount

PeerToPeerWithMetadata

Preburn

PublishSharedEd2551PublicKey

RegisterValidatorConfig

RemoveValidatorAndReconfigure

RotateAuthenticationKey

RotateAuthenticationKeyWithNonce

RotateAuthenticationKeyWithNonceAdmin

RotateAuthenticationKeyWithRecoveryAddress

RotateDualAttestationInfo

RotateSharedEd2551PublicKey

SetValidatorConfigAndReconfigure

SetValidatorOperator

SetValidatorOperatorWithNonceAdmin

TieredMint

UnfreezeAccount

UpdateExchangeRate

UpdateDiemVersion

UpdateMintingAbility

UpdateDualAttestationLimit

Implementations

impl LegacyStdlibScript

pub fn all() -> Vec<Self>

Return a vector containing all of the standard library scripts (i.e., all inhabitants of the StdlibScript enum)

pub fn allowlist() -> Vec<HashValue>

Construct the allowlist of script hashes used to determine whether a transaction script can be executed on the Diem blockchain

pub fn name(self) -> String

Return a lowercase-underscore style name for this script

pub fn is(code_bytes: &[u8]) -> bool

Return true if code_bytes is the bytecode of one of the standard library scripts

pub fn compiled_bytes(self) -> CompiledBytes

Return the Move bytecode that was produced by compiling this script.

pub fn abi(self) -> TransactionScriptABI

Return the ABI of the script (including the bytecode). NB: For legacy scripts, the script ABI will always be a TransactionScriptABI.

pub fn hash(self) -> HashValue

Return the sha3-256 hash of the compiled script bytes.